#1381 - Modularize CasDiff code
- Move CasDiff inner classes into their own respective source files
