+++ webanno-curation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/curation/casdiff/api/CasDiffException.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.clarin.webanno.curation.casdiff;
+package de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.api;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.exception.AnnotationException;
