+++ webanno-curation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/curation/casdiff/api/DiffAdapter.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2019
+ * Ubiquitous Knowledge Processing (UKP) Lab and FG Language Technology
+ * Technische Universität Darmstadt
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.api;
+
+import java.util.Collection;
+import java.util.Set;
+
+import org.apache.uima.cas.FeatureStructure;
+import org.apache.uima.cas.text.AnnotationFS;
+
+import de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.LinkCompareBehavior;
+import de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.LinkFeatureDecl;
+
+public interface DiffAdapter
+{
+    String getType();
+    
+    Collection<? extends Position> generateSubPositions(int aCasId, AnnotationFS aFs,
+            LinkCompareBehavior aLinkCompareBehavior);
+
+    LinkFeatureDecl getLinkFeature(String aFeature);
+    
+    Set<String> getLabelFeatures();
+    
+    Position getPosition(int aCasId, FeatureStructure aFS);
+
+    Position getPosition(int aCasId, FeatureStructure aFS, String aFeature, String aRole,
+            int aLinkTargetBegin, int aLinkTargetEnd, LinkCompareBehavior aLinkCompareBehavior);
+}
